// InterfacesConsolidadas/Web/src/app/maestros/colonias/models/coloniaSad.model.ts
export class ColoniaSad {
  id: number;
  sucursalId: number;
  sucursal: string;
  coloniaId: number;
  colonia: string;
  horaAperturaLunesAviernes: string;
  horaCierreLunesAviernes: string;
  horaAperturaSabados: string;
  horaCierreSabados: string;
  horaAperturaDomingos: string;
  horaCierreDomingos: string;
  activo: boolean;
  ciudad: string;
  ciudadId: string;
  seleccionada: boolean;
}